[loader] menu.c: Allow "going back" to device selection after selecting SD or USB.
This allows the user to switch between SD and USB without having to reload
the entire application.

TODO: If SD or USB is missing and/or doesn't have a games directory, the
program will exit.

ShowLoadingScreen(): New function that consolidates all of the "Loading"
screen invocations and makes it easier to update.