[loader] menu.c: Don't attempt to use the PARTITION* from devoptab_list[] if the device wasn't opened.
If the device wasn't opened, deviceData contains random garbage, and
attempting to read it crashes the program.

This fixes a crash that occurred when I attempted to use my test SD card
partitioned as GPT and containing 4 1MB empty partitions and 1 FAT32
partition. This branch doesn't have the improved GPT code, so it only
checked the first four partitions before giving up.