[kernel] Use multiple volume support in FatFS.
This replaces the hackish function pointer usage in diskio.c with
correct pdrv usage, similar to the loader diskio.c in the
fatfsloader branch.

Note that this increases the code size a bit, so I increased the
code segment by 32 KB, which takes away 32 KB from the data segment.

main.c: f_chdrive() to the correct drive after mounting. This probably
isn't necessary, since FatFS appears to do this itself, but we might
as well change the drive just to be sure.

TODO:
- Get s_cnt from the SD card CSD instead of the FAT filesystem.
- Don't set s_size/s_cnt in usbstorage.c. Instead, provide a function
  to retrieve the values, then use that function in disk_initialize().