[loader] MemCard.c: Optimized memory card formatting.
Use a struct for the header and direct pointer access for the directory
and block tables. The memory card data is now allocated using memalign(),
so all of the fields are guaranteed to be aligned correctly.

TODO: Set up header->reserved1[] and header->formatTime correctly
instead of just setting them to 0?