[kernel] EXI.c: Initial attempt at adding support for Slot B memory cards.
If a file named GSNE_B.raw (for SA2B) or ninmemb.raw/ninmemjb.raw (for
MULTI card) exists, and the total size of both the Slot A image and
the Slot B image is less than 16 MB, Slot B will be enabled.

...or *should* be enabled. SA2B isn't recognizing that a second card
exists. There might be a missing patch somewhere.