[kernel] EXIProbe.S: Add support for Slot B.
Thanks to FIX94 for pointing out EXIProbe.S as the reason why
games weren't detecting a card in Slot B.

CommonConfig.h: Added an internal configuration bit indicating
if Slot B is active. This is used by EXIProbe.S to determine if
it should report success or failure.

FIXME: SA2B now reports that there's "something" in Slot B,
but claims the card is damaged and cannot be used.