[loader] Adjust the background image for widescreen.
If the system's widescreen setting is enabled, scale the background image
to 480x480 and draw it in the center of the screen. This makes it look
correct on widescreen displays. Also, fill the now-unused area with the
same color as the image's primary background color.

Optimized background.png to take up slightly less space.
(91,149 -> 90,813)