PADReadGC.c: Reduce complexity in the HID D-Pad code.
We don't need to check if Up, UpLeft, or UpRight is pressed before
setting PAD_BUTTON_UP. Instead, set PAD_BUTTON_UP if Up is pressed.
If UpLeft is pressed, then set PAD_BUTTON_UP | PAD_BUTTON_LEFT.
Repeat for UpRight and all the other directions.

The result is effectively the same, but with fewer branches and a
smaller code size: 9,272 -> 9,160