[loader] usbstorage.c: Check for a valid MBR signature before accepting a device.
This fixes the scenario where a user has both a Wii U-formatted drive
and a Wii/GCN FAT32 drive connected at the same time.

Tested on Wii using a Wii U-formatted USB device on port 1 and a FAT32
drive on port 0. Before this patch, Nintendont complained that it
couldn't find the USB FAT device. With this patch, it works correctly.