Enable GameCube controllers on Wii U (vWii).
GameCube controllers *normally* cannot be connected to a Wii U directly.

The CAT-DEV Wii U devkit has a Classic Controller port that goes to a
device that converts the CC protocol to GameCube SI protocol, which
then goes to test point TP69 on the system board. It turns out that
test point TP69 exists on retail systems, so it's possible to wire up
a single GameCube port on any Wii U.

This commit removes the SI restrictions on vWii and adds a note about
using TP69 in the "Native Control" option. (It works with the option
set to either On or Off.)